Add files via upload
This program i've written for the final project on Coursera for the OS interaction course. In it is python code to open syslog, iterate through the file, use regex to take in ERROR, INFO and Usernames and put them into a list. The list is stored in a csv file which then converted into html file using bash commands.